requirements.txt
wheel
typing-extensions
hopsworks
pandas
requests
scikit-learn
modal-client
python-dotenv
dataframe_image
